+++ lib/main.dart
@@ -3,6 +3,7 @@ import 'package:get/get.dart';
 import 'package:mentalwaev/theme/theme.dart';
 import 'package:mentalwaev/views/auth/login.dart';
 import 'package:mentalwaev/views/home/home.dart';
+import 'package:mentalwaev/views/meditation/new_meditation_screen.dart';
 import 'package:mentalwaev/views/timer/timer_screen.dart';
 
 void main() {
@@ -18,13 +19,13 @@ class MyApp extends StatelessWidget {
     return GetMaterialApp(
       title: 'Flutter Demo',
       theme: AppTheme().themeData,
-      initialRoute: "/login",
+      initialRoute: "/newMeditation",
       debugShowCheckedModeBanner: false,
       routes: {
         '/home': (context) => const Home(),
         '/login': (context) => const SignUp(),
         '/timerScreen': (context) => const MeditationTimerScreen(),
-
+        '/newMeditation': (context) => const NewMeditationScreen(),
       },
     );
   }
