--- lib/views/meditation/motivation_blo.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MotivationBla extends StatefulWidget {
-  const MotivationBla({Key? key}) : super(key: key);
-
-  @override
-  State<MotivationBla> createState() => _MotivationBlaState();
-}
-
-class _MotivationBlaState extends State<MotivationBla> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold();
-  }
-}
