+++ lib/widgets/meditation_preset_card.dart
@@ -6,6 +6,7 @@ class MeditationPresetCard extends StatelessWidget {
   final Color color;
   final double leftPadding;
   final double rightPadding;
+  final VoidCallback onTap;
   const MeditationPresetCard({
     Key? key,
     required this.prefix,
@@ -13,23 +14,27 @@ class MeditationPresetCard extends StatelessWidget {
     required this.color,
     required this.leftPadding,
     required this.rightPadding,
+    required this.onTap,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      padding: EdgeInsets.fromLTRB(leftPadding, 17.0, rightPadding, 17.0),
-      decoration: BoxDecoration(
-        color: color,
-        borderRadius: BorderRadius.circular(10.0),
-      ),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          prefix,
-          suffix,
-        ],
+    return InkWell(
+      onTap: onTap,
+      child: Container(
+        padding: EdgeInsets.fromLTRB(leftPadding, 17.0, rightPadding, 17.0),
+        decoration: BoxDecoration(
+          color: color,
+          borderRadius: BorderRadius.circular(10.0),
+        ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            prefix,
+            suffix,
+          ],
+        ),
       ),
     );
   }
